third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-dom-markerUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'markerUnits' attribute of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.getAttribute('markerUnits') is "userSpaceOnUse"
PASS markerElement.getAttribute('markerUnits') is "strokeWidth"
PASS successfullyParsed is true

TEST COMPLETE
